backend/requirements.txt
Django
djangorestframework
langchain
python-dotenv
langchain-groq
langchain-community
crewai



